+++ src/main/java/com/equip/equiprental/rental/repository/RentalRepository.java
@@ -4,5 +4,5 @@ import com.equip.equiprental.rental.domain.Rental;
 import com.equip.equiprental.rental.repository.dsl.RentalQRepo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Integer>, RentalQRepo {
+public interface RentalRepository extends JpaRepository<Rental, Long>, RentalQRepo {
 }
